# Conflicts:
#	Villain CES/Villain CES/Entity.h
#	Villain CES/Villain CES/GameManger.cpp
#	Villain CES/Villain CES/InputSystem.cpp
#	Villain CES/Villain CES/MyPixelShader.csh
#	Villain CES/Villain CES/MyVertexShader.csh

+++ Villain CES/Villain CES/ProjectileSystem.cpp
@@ -29,7 +29,7 @@ int CProjectileSystem::CreateBulletProjectile(int indexofBullet,Clips* Gun)
 bool CProjectileSystem::Reload(Clips * Gun)
 {
 	if (Gun->nBulletsAvailables.size() != Gun->nSizeofClipl) {
-		for (int i = Gun->nBulletsAvailables.size(); i < Gun->nSizeofClipl; ++i) {
+		for (size_t i = Gun->nBulletsAvailables.size(); i < Gun->nSizeofClipl; ++i) {
 			Gun->nBulletsAvailables.push_back(true);
 			
 		}
